lambda_functions/inference/requirements.txt
boto3>=1.26.0
botocore>=1.29.0
shared
strands-agents
strands-agents-tools
strands-agents-builder